// src/enums/enums.ts
export enum AlertSiteTypes {
    error = 'error',
    warning = 'warning',
    info = 'info',
    success = 'success'
}

export enum cartProductType {
    album = 'Album',
    albumFile = 'AlbumFile',
    poster = 'Poster',
    book = 'Book'
}

export enum UserLoginTypes {
    signIn = 'sign-in',
    signUp = 'sign-up',
    forgotPassword = 'forgot-password'
}

export enum OrderItemTypes {
    AlbumFile = 'AlbumFile',
    Poster = 'Poster',
    Book = 'Book',
    Album = 'Album'
}